fix: reach zero javac warnings and refresh CI actions
- Enable -Xlint:all except this-escape (documented NeoForge registration
  pattern); all remaining categories are fixed at the source.
- Replace deprecated hasChunkAt(BlockPos) with hasChunk on section coords,
  Abilities.mayfly reads with player.mayFly(), TextureAtlas.LOCATION_BLOCKS
  with InventoryMenu.BLOCK_ATLAS, deprecated defineListAllowEmpty overload,
  and move the condensation matrix cutout render type into its generated
  model instead of ItemBlockRenderTypes.
- Remove redundant casts, one rawtype and add scoped unchecked suppressions
  where AE2's generic API forces erasure.
- Bump GitHub Actions to current majors (checkout v7, setup-java v6,
  setup-gradle v6, upload-artifact v7, setup-python v7, gh-release v3) to
  clear the Node 20 deprecation annotations.

+++ src/main/java/com/raishxn/ufo/block/entity/DimensionalMatterAssemblerBlockEntity.java
@@ -755,7 +755,7 @@ public class DimensionalMatterAssemblerBlockEntity extends AENetworkedPoweredBlo
             for (var req : recipe.getItemInputs()) {
                 if (req == null || req.isEmpty())
                     continue;
-                int amountNeeded = (int) req.getAmount();
+                int amountNeeded = req.getAmount();
                 for (var stack : availableInputs) {
                     if (req.getIngredient().test(stack)) {
                         int toTake = Math.min(stack.getCount(), amountNeeded);
@@ -933,7 +933,7 @@ public class DimensionalMatterAssemblerBlockEntity extends AENetworkedPoweredBlo
 
                     if (factor >= 1) {
                         var extracted = src.extractAEPower(
-                                (double) powerConsumption * factor / speedFactor,
+                                powerConsumption * factor / speedFactor,
                                 Actionable.MODULATE,
                                 PowerMultiplier.CONFIG);
                         var actualFactor = (int) Math.floor(extracted / powerConsumption * speedFactor);
@@ -984,7 +984,7 @@ public class DimensionalMatterAssemblerBlockEntity extends AENetworkedPoweredBlo
                     // Consume inputs shapelessly
                     for (var req : out.getItemInputs()) {
                         if (req != null && !req.isEmpty()) {
-                            int amountNeeded = (int) req.getAmount();
+                            int amountNeeded = req.getAmount();
                             for (int i = 0; i < this.inputInv.size() && amountNeeded > 0; i++) {
                                 var currentStack = this.inputInv.getStackInSlot(i);
                                 if (req.getIngredient().test(currentStack)) {
